backend/requirements.txt
# FastAPI 및 서버
fastapi>=0.115.0  # Python 3.13 호환
uvicorn[standard]>=0.32.0
pydantic>=2.10.0  # Python 3.13 호환
pydantic-settings>=2.6.0

# 데이터 수집 및 처리
yfinance==0.2.35
pandas>=2.2.0  # Python 3.13 호환
numpy>=1.26.0
finance-datareader  # 한국 주식 데이터
pykrx  # 한국거래소 데이터

# AI 모델 API 클라이언트
openai==1.10.0
anthropic==0.8.1
google-cloud-aiplatform==1.40.0
google-generativeai==0.3.2
httpx==0.26.0
aiohttp==3.9.1

# 뉴스 및 감성 분석
requests==2.31.0
beautifulsoup4==4.12.3
textblob==0.17.1

# 캐싱 및 저장
redis==5.0.1
boto3==1.34.34  # S3 사용 시

# 유틸리티
python-dotenv==1.0.0
python-multipart==0.0.6

# 테스트
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
hypothesis==6.92.2
httpx==0.26.0  # FastAPI 테스트용

# 로깅 및 모니터링
loguru==0.7.2
